Move bundles to new file
