+++ src/prelude.rs
@@ -1,7 +1,7 @@
 pub use crate::components::audio_listener::AudioListener;
 pub use crate::components::audio_source::AudioSource;
+pub use crate::components::bundles::SpatialAudioBundle;
+pub use crate::components::bundles::SpatialListenerBundle;
 pub use crate::components::velocity::Velocity;
-pub use crate::components::SpatialAudioBundle;
-pub use crate::components::SpatialListenerBundle;
 pub use crate::fmod_plugin::FmodPlugin;
 pub use crate::fmod_studio::FmodStudio;
